requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
faiss-cpu>=1.7.2  # I'm using iMac so using CPU, if you have dedicated GPU then use faiss-gpu
Pillow>=8.0.0
